diabetes_agent/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
chromadb==0.4.22
sentence-transformers==2.3.1
pydantic==2.5.3
transformers==4.37.2
torch==2.2.0
Pillow==10.2.0
python-multipart==0.0.6
